AI/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
psycopg2-binary==2.9.9
python-decouple==3.8
